netx: merge public API with internal
Closes https://github.com/ooni/probe-engine/issues/338, because I have
added code and unit test to check for such case.

Closes https://github.com/ooni/probe-engine/issues/310.

Part of https://github.com/ooni/probe-engine/issues/302.
