+++ internal/netxlogger/netxlogger_test.go
@@ -2,18 +2,29 @@ package netxlogger
 
 import (
 	"io/ioutil"
+	"net/http"
 	"testing"
+	"time"
 
 	"github.com/apex/log"
 	"github.com/apex/log/handlers/discard"
 	"github.com/ooni/probe-engine/netx"
+	"github.com/ooni/probe-engine/netx/modelx"
 )
 
 func TestIntegration(t *testing.T) {
 	log.SetHandler(discard.Default)
-	client := netx.NewHTTPClient(NewHandler(log.Log))
+	client := netx.NewHTTPClient()
 	client.ConfigureDNS("udp", "dns.google.com:53")
-	resp, err := client.HTTPClient.Get("http://www.facebook.com")
+	req, err := http.NewRequest("GET", "http://www.facebook.com", nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+	req = req.WithContext(modelx.WithMeasurementRoot(req.Context(), &modelx.MeasurementRoot{
+		Beginning: time.Now(),
+		Handler:   NewHandler(log.Log),
+	}))
+	resp, err := client.HTTPClient.Do(req)
 	if err != nil {
 		t.Fatal(err)
 	}
